+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- [#1668](https://github.com/teambit/bit/issues/1668) bug fix - bit watch doesn't update files
+- [#1668](https://github.com/teambit/bit/issues/1668) bug fix - `bit watch` doesn't update files
+- improve `bit watch` to watch directories instead of only files to support addition / deletion
 
 ## [14.2.1] - 2019-07-21